Fix for PostProccessing, Python3 compatibility and some new tests (#3221)
* Implement some tests of config_tests.py

Notes:
check_setting_int + check_setting_float have unreachable code when value is of None type.
They don't get to check str(my_val) == str(None) because an exception is already thrown when float/int(config[cfg_name][item_name]) fails to convert NoneType.

* ConfigTestChanges

* ConfigTestChanges, try #2

* ConfigTestChanges, now it'll work.

* The best I could do right now for config_tests.py

* Add more checks in config.check_setting_*
Add missing tzlocal to libs,
Rename minimax to mini_max and CheckSection to check_section
sixify sickbeard.config and use unicode_literals

* Updated migration names.

* Correct Config Migration v9 labels

* See how converting to six works out -DNM

* More unicode_literals, six conversions, and fixes

* Update concurrent.futures

* More unicode_literals, six conversions, and fixes

* Remove u prefixes, add unicode_literals

* Fix index errors for sqlite3.Row, fix unquote_plus call/import

* No need to check str(None) since it will already throw. You were right =P

* An attempt to fix #3231, fix #3230, fix #3220, fix #3216, and fix #3052

* HelpersZipTests

* More sqlite3.Row fixes

* Fix up subtitle renaming in the post processor, Fix error trying to replace_extension on the cur_file_name (basename) rather than the cur_file_path (full name with extension) <-- This might have been the whole pp problem.

* More sqlite3.Row, fix remaining pp issues

* Small adjustment

* /slap

* Fix endswith

+++ sickbeard/common.py
@@ -33,6 +33,8 @@ import uuid
 from os import path
 
 from fake_useragent import UserAgent, settings as UA_SETTINGS
+import six
+from six.moves import reduce
 
 from sickbeard.numdict import NumDict
 from sickrage.helper import video_screen_size
@@ -431,7 +433,7 @@ class Quality(object):
         :param status: to split
         :returns: a tuple containing (status, quality)
         """
-        status = long(status)
+        status = int(status)
         if status == UNKNOWN:
             return UNKNOWN, Quality.UNKNOWN
